+++ src/models/AdjacencyList/AdjacencyList.ts
@@ -1,5 +1,6 @@
 import { Vertex } from '../Vertex/Vertex/Vertex';
 import { MutableHashMap } from '@tgillespie/hash-data-structures';
+import {Edge} from "../Edge";
 
 export class AdjacencyList<V extends Vertex> {
 	protected _adjacencyList: MutableHashMap<V, V[]>;
@@ -31,5 +32,14 @@ export class AdjacencyList<V extends Vertex> {
         for(const [key, value] of this._adjacencyList.entries()) {
             this._adjacencyList.set(key, value.filter(x => !x.equals(vertex)));
         }
+    }
+
+    deleteEdge(edge: Edge<V>): void {
+        const previousDirectedEdgeVertices = this._adjacencyList.get(edge.vertexA);
+        this._adjacencyList.set(edge.vertexA, previousDirectedEdgeVertices.filter(x => !x.equals(edge.vertexB)));
+
+        // Because it is about adjacency it must be done for both directed and undirected edges
+        const previousOppositeEdge = this._adjacencyList.get(edge.vertexB);
+        this._adjacencyList.set(edge.vertexB, previousOppositeEdge.filter(x => !x.equals(edge.vertexA)));
     }
 }